Fixed birthday cake, moved files, renamed files
